+++ R/api_00.R
@@ -6,7 +6,8 @@ ua <- httr::user_agent("http://github.com/tychobra/polished")
 #'
 #' @return an S3 object of class "polished_api_res".
 #'
-#' @importFrom httr http_type
+#' @importFrom httr content http_error http_type status_code
+#' @importFrom jsonlite fromJSON
 #'
 #' @export
 #'
@@ -55,13 +56,14 @@ print.polished_api_res <- function(x, ...) {
 
 #' set Polished API key
 #'
-#' The API key is set as an R option at \code{getOption("polished")$api_key}.
+#' The API key can be set as an Environment Variable via
+#' \code{Sys.getenv("POLISHED_API_KEY")}.
 #'
 #' @param api_key the Polished API key
 #'
 #' @export
 #'
-#' @return a list of the newly set polished R options
+#' @return a list of the newly set `polished` R options
 #'
 #' @examples
 #'
@@ -70,43 +72,41 @@ print.polished_api_res <- function(x, ...) {
 #'
 set_api_key <- function(api_key) {
 
-  current_polished_options <- getOption("polished")
+  assign("api_key", api_key, envir = .polished)
 
-  if (is.null(current_polished_options)) {
-    out <- list(
-      api_key = api_key
-    )
-  } else {
-    out <- current_polished_options
-    out$api_key <- api_key
-  }
-
-  options("polished" = out)
-
-  invisible(out)
+  invisible(api_key)
 }
+
+#' @export
+#' @rdname set_api_key
+get_api_key <- function() {
+
+  api_key <- .polished$api_key
+
+
+  if (is.null(api_key)) {
+    api_key <- Sys.getenv("POLISHED_API_KEY", unset = NA)
+    if (is.na(api_key)) {
+      stop("polished API key must be set", call. = FALSE)
+    }
+  }
+  api_key
+}
+
 
 
 set_api_url <- function(
   api_url = "https://auth-api.polished.tech/v1",
   host_api_url = "https://host-api.polished.tech/v1"
 ) {
-  current_polished_options <- getOption("polished")
 
-  if (is.null(current_polished_options)) {
-    out <- list(
-      api_url = api_url,
-      host_api_url = host_api_url
-    )
-  } else {
-    out <- current_polished_options
-    out$api_url <- api_url
-    out$host_api_url <- host_api_url
-  }
+  assign("api_url", api_url, envir = .polished)
+  assign("host_api_url", host_api_url, envir = .polished)
 
-  options("polished" = out)
-
-  invisible(out)
+  invisible(list(
+    api_url = api_url,
+    host_api_url = host_api_url
+  ))
 }