+++ R/api_02_users.R
@@ -12,16 +12,16 @@
 #' and the \code{email} will be ignored.  If the user does not exists, a zero row tibble
 #' will be returned.
 #'
-#' @return an object of class \code{polished_api_res}.  The "content" of the object is a
+#' @return an object of class \code{polished_api_res}.  The `content` of the object is a
 #' tibble of users(s) with the following columns:
-#' - uid
-#' - email
-#' - email_verified
-#' - created_by
-#' - created_at
-#' - modified_by
-#' - modified_at
-#' - is_password_set
+#' - `uid`
+#' - `email`
+#' - `email_verified`
+#' - `created_by`
+#' - `created_at`
+#' - `modified_by`
+#' - `modified_at`
+#' - `is_password_set`
 #'
 #' @export
 #'
@@ -32,7 +32,7 @@
 get_users <- function(
   user_uid = NULL,
   email = NULL,
-  api_key = getOption("polished")$api_key
+  api_key = get_api_key()
 ) {
 
   query_out <- list()
@@ -40,7 +40,7 @@ get_users <- function(
   query_out$email <- email
 
   resp <- httr::GET(
-    url = paste0(getOption("polished")$api_url, "/users"),
+    url = paste0(.polished$api_url, "/users"),
     ua,
     httr::authenticate(
       user = api_key,
@@ -59,7 +59,7 @@ get_users <- function(
 
 #' Polished API - Add a User
 #'
-#' @param email an email address.
+#' @param email the new user's email address.
 #'
 #' @inheritParams get_apps
 #'
@@ -69,14 +69,14 @@ get_users <- function(
 #'
 #' @importFrom httr POST authenticate
 #'
-add_user <- function(email, api_key = getOption("polished")$api_key) {
+add_user <- function(email, api_key = get_api_key()) {
 
   body_out <- list(
     email = email
   )
 
   resp <- httr::POST(
-    url = paste0(getOption("polished")$api_url, "/users"),
+    url = paste0(.polished$api_url, "/users"),
     ua,
     httr::authenticate(
       user = api_key,
@@ -94,7 +94,7 @@ add_user <- function(email, api_key = getOption("polished")$api_key) {
 
 #' Polished API - Delete a User
 #'
-#' @param user_uid the user uid of the user to be deleted.
+#' @param user_uid the uid of the user to be deleted.
 #'
 #' @inheritParams get_apps
 #'
@@ -104,14 +104,14 @@ add_user <- function(email, api_key = getOption("polished")$api_key) {
 #'
 #' @importFrom httr DELETE authenticate
 #'
-delete_user <- function(user_uid, api_key = getOption("polished")$api_key) {
+delete_user <- function(user_uid, api_key = get_api_key()) {
 
   query_out <- list(
     user_uid = user_uid
   )
 
   resp <- httr::DELETE(
-    url = paste0(getOption("polished")$api_url, "/users"),
+    url = paste0(.polished$api_url, "/users"),
     ua,
     httr::authenticate(
       user = api_key,