+++ R/package_deps.R
@@ -11,8 +11,8 @@
 #'   validating that each detection is indeed a valid `CRAN` or public `GitHub` package
 #'   and can be installed.
 #'
-#' @param app_dir path to a directory containing R scripts or RMarkdown files. Defaults to current working directory if left blank.
-#' @param verbose logical - defaults to TRUE. Will provide feedback to detected or invalid R packages.
+#' @param app_dir path to a directory containing R scripts or R Markdown files. Defaults to current working directory if left blank.
+#' @param verbose boolean (default: \code{TRUE}) - Provide feedback about detected invalid R packages.
 #'
 #' @return a list of package dependencies
 #'