+++ R/secure_static.R
@@ -5,7 +5,7 @@
 #' and flexdashboards.
 #'
 #' @param html_file_path the path the to HTML file.  See the details for more info.
-#' @param global_sessions_config_args arguments to be passed to \code{\link{global_sessions_config}}.
+#' @param polished_config_args arguments to be passed to \code{\link{polished_config}}.
 #' @param sign_out_button action button or link with \code{inputId = "sign_out"}. Set to \code{NULL} to not include a sign out button.
 #'
 #' @md
@@ -24,12 +24,12 @@
 #'
 #' @return a Shiny app object
 #'
-#' @importFrom shiny shinyApp actionLink
+#' @importFrom shiny shinyApp actionButton actionLink icon observeEvent
 #' @importFrom htmltools tags tagList
 #'
 secure_static <- function(
   html_file_path,
-  global_sessions_config_args,
+  polished_config_args,
   sign_out_button = shiny::actionLink(
     "sign_out",
     "Sign Out",
@@ -80,7 +80,7 @@ secure_static <- function(
 
   server <- secure_server(function(input, output, session) {
 
-    observeEvent(input$sign_out, {
+    shiny::observeEvent(input$sign_out, {
 
       tryCatch({
         sign_out_from_shiny(session)
@@ -98,8 +98,8 @@ secure_static <- function(
 
     # configure the global sessions when the app initially starts up.
     do.call(
-      "global_sessions_config",
-      global_sessions_config_args
+      "polished_config",
+      polished_config_args
     )
   })
 }