+++ R/send_password_reset_email_module.R
@@ -2,12 +2,13 @@
 #'
 #' @param id the Shiny module \code{id}
 #'
-#' @importFrom shiny actionLink
+#' @importFrom htmltools tagList
+#' @importFrom shiny actionLink NS
 #' @importFrom shinyFeedback useShinyFeedback
 #'
 #' @export
 send_password_reset_email_module_ui <- function(id) {
-  ns <- NS(id)
+  ns <- shiny::NS(id)
 
   tagList(
     shinyFeedback::useShinyFeedback(feedback = FALSE),
@@ -30,7 +31,7 @@ send_password_reset_email_module_ui <- function(id) {
 #' reset email to.
 #'
 #' @importFrom shiny observeEvent
-#' @importFrom httr POST authenticate status_code
+#' @importFrom httr POST authenticate content status_code
 #' @importFrom jsonlite fromJSON
 #' @importFrom shinyFeedback showToast
 #'
@@ -44,15 +45,15 @@ send_password_reset_email_module <- function(input, output, session, email) {
 
     tryCatch({
       res <- httr::POST(
-        url = paste0(getOption("polished")$api_url, "/send-password-reset-email"),
+        url = paste0(.polished$api_url, "/send-password-reset-email"),
         httr::authenticate(
-          user = getOption("polished")$api_key,
+          user = get_api_key(),
           password = ""
         ),
         body = list(
           email = hold_email,
-          app_uid = getOption("polished")$app_uid,
-          is_invite_required = .global_sessions$is_invite_required
+          app_uid = .polished$app_uid,
+          is_invite_required = .polished$is_invite_required
         ),
         encode = "json"
       )