+++ inst/examples/polished_example_01/server.R
@@ -1,20 +1,24 @@
 server <- function(input, output, session) {
 
   output$secure_content <- renderPrint({
-    session$userData$user()
+    session$user
   })
 
 
   observeEvent(input$sign_out, {
 
-    sign_out_from_shiny(session)
-    session$reload()
+    tryCatch({
 
-  })
+      sign_out_from_shiny(session)
+      session$reload()
 
-  output$test_render <- renderUI({
+    }, error = function(err) {
 
-    session$userData$user()
+      msg <- "unable to sign out"
+      print(msg)
+      print(err)
+
+    })
 
   })